backend/requirements.txt
google-adk>=1.0.0
a2a-sdk>=0.2.0
ag-ui-protocol>=0.1.14
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
